update intensity transform
add gaussian noise
