+++ pymic/transform/trans_dict.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 from __future__ import print_function, division
-from pymic.transform.gamma_correction import  ChannelWiseGammaCorrection
+from pymic.transform.intensity import  *
 from pymic.transform.gray2rgb import GrayscaleToRGB
 from pymic.transform.flip import RandomFlip
+from pymic.transform.intensity import GaussianNoise
 from pymic.transform.pad import Pad
 from pymic.transform.rotate import RandomRotate
 from pymic.transform.rescale import Rescale, RandomRescale  
@@ -12,12 +13,13 @@ from pymic.transform.crop import *
 from pymic.transform.label_convert import * 
 
 TransformDict = {
-    'ChannelWiseGammaCorrection': ChannelWiseGammaCorrection,
     'ChannelWiseThreshold': ChannelWiseThreshold,
     'ChannelWiseThresholdWithNormalize': ChannelWiseThresholdWithNormalize,
     'CropWithBoundingBox': CropWithBoundingBox,
     'CenterCrop': CenterCrop,
     'GrayscaleToRGB': GrayscaleToRGB,
+    'GammaCorrection': GammaCorrection,
+    'GaussianNoise': GaussianNoise,
     'LabelConvert': LabelConvert,
     'LabelConvertNonzero': LabelConvertNonzero,
     'LabelToProbability': LabelToProbability,
